FEAT: Implementa o uso correto de decimais e Model Binding pt-BR
Este commit resolve vários problemas que impediam o cadastro correto de produtos com valores decimais (vírgula).

Alterações principais:
Configuração da aplicação para a cultura 'pt-BR' para interpretar vírgula (,) como separador decimal.
Refatoração do ProdutoModel no C# para usar 'decimal' e propriedades {get; set;}.
Correção da View Razor (CadastrarProduto.cshtml) para usar asp-for e passar os valores corretos do Enum.

+++ REPOSITORY/Produto/ProdutoRepository.cs
@@ -97,7 +97,7 @@ namespace REPOSITORY.Produto
                     ProdutoModel produto = new() {
                         CodigoProduto = Reader.GetInt32(CodigoOrdinal),
                         NomeProduto = Reader.GetString(NomeOrdinal),
-                        ValorProduto = Reader.GetDouble(ValorOrdinal),
+                        ValorProduto = Reader.GetDecimal(ValorOrdinal),
                         Disponibilidade = (DisponibilidadeEnum)Reader.GetInt32(DisponibilidadeOrdinal)
                     };
                     ListaProdutos.Add(produto);
